Move s_aim_assist_targetting_data into aim_assist header
